+++ cogs/events.py
@@ -3,9 +3,9 @@ import asyncio
 import re
 
 from discord.ext import commands
-from db.dbase import DBase
 import discord
 
+from db.dbase import DBase
 from cogs.utils.messages import delete_all, MessageManager
 from cogs.utils.checks import is_event, is_admin, is_int
 from cogs.utils import constants