+++ cogs/settings.py
@@ -1,9 +1,9 @@
 import asyncio
 
 from discord.ext import commands
-from db.dbase import DBase
 import discord
 
+from db.dbase import DBase
 from cogs.utils.messages import MessageManager
 from cogs.utils.checks import is_admin