+++ cogs/general.py
@@ -1,13 +1,37 @@
+from datetime import datetime
+
 import discord
 from discord.ext import commands
 
 from cogs.utils.messages import MessageManager
+from cogs.utils import constants
 
 
-class Misc:
+class General:
 
     def __init__(self, bot):
         self.bot = bot
+
+
+    @commands.command(pass_context=True)
+    async def countdown(self, ctx):
+        """Show time until upcoming Destiny 2 releases"""
+        manager = MessageManager(self.bot, ctx.message.author, ctx.message.channel, [ctx.message])
+
+        text = ""
+        for name, date in constants.RELEASE_DATES:
+            diff = date - datetime.now()
+            if diff.days == -1:
+                text += "{}: Today!\n".format(name)
+            elif diff.days == 0:
+                text += "{}: Tomorrow!\n".format(name)
+            elif diff.days > 1:
+                text += "{}: {} days\n".format(name, diff.days)
+
+        countdown = discord.Embed(title="Destiny 2 Countdown", color=constants.BLUE)
+        countdown.description = text
+        await manager.say(countdown, embed=True, delete=False)
+        await manager.clear()
 
 
     @commands.command(pass_context=True)