+++ cogs/help.py
@@ -68,7 +68,7 @@ class Help:
             help = discord.Embed(title="Available Commands", color=constants.BLUE)
             help.description = ("Items in <angled_brackets> are *required*"
                               + "\nItems in [square_brackets] are *optional*"
-                              + "\nThe [] and <> characters are to be omitted")
+                              + "\nNote - don't include the [] and <> characters")
             help.set_footer(text="Use {}help [command] for more info on a command".format(prefix))
 
             for key in commands: