+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Help command for individual commands (Ex. !help event)
+### Changed
+- Help command now displays command arguments
+
 ## [0.4.0] - 2017-08-08
 ### Added
 - Users can now add their timezone to the roster with the new 'timezone' command