+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.5.0] - 2017-08-12
 ### Added
+- New command 'countdown' displays time until upcoming Destiny 2 releases
 - Help command now supports individual commands and subcommands (Eg. '!help event')
 - Command help messages are now much more descriptive
 - Error message now displays when an invalid command is invoked, or an argument is missing